Function tp0::inicializar::cargar_juego

pub fn cargar_juego(
    maze_file_name: &str,
    output_file: &mut File
) -> Result<Juego>

Carga el estado del juego desde un archivo de laberinto especificado y lo inicializa.

Argumentos

maze_file_name: El nombre del archivo de laberinto desde el cual se cargará el juego.

output_file: Archivo de salida donde se escribirán mensajes de error si es necesario.

Devuelve

Retorna un Result que contiene el objeto Juego inicializado si la carga y procesamiento del archivo fueron exitosos, o un error en caso contrario.

Esta función realiza las siguientes acciones:

Abre el archivo de laberinto especificado.

Lee cada línea del archivo, procesando la configuración del juego y actualizando el estado de juego.

Maneja los errores de lectura y procesamiento, escribiendo mensajes en output_file y devolviendo errores apropiados.

Verifica si el tablero ingresado es cuadrado.

Inicializa la dimensión del juego en base al número de filas procesadas.

Si todo se realiza correctamente, devuelve Ok(juego). En caso de errores, se devuelve un error io::Result.