Struct tp0::juego::Juego

pub struct Juego {
    dimension: i8,
    enemigos: Vec<Enemigo>,
    obstaculos: Vec<Obstaculo>,
    bombas: Vec<Bomba>,
    desvios: Vec<Desvio>,
}

Fields
dimension: i8enemigos: Vec<Enemigo>obstaculos: Vec<Obstaculo>bombas: Vec<Bomba>desvios: Vec<Desvio>
Implementations

impl Juego

pub fn new() -> Juego

pub fn inicializar_dimension(&mut self, dimension: i8)

Inicializa la dimensión del juego.

Esta función establece la dimensión del juego con el valor proporcionado.

Argumentos

dimension: El valor de la dimensión del juego.

pub fn inicializar_desvio( &mut self, coordenada: Coordenada, direccion: String, id: String )

Inicializa un desvío en el juego.

Esta función crea y agrega un nuevo desvío al juego con la coordenada, dirección e identificador especificados.

Argumentos

coordenada: La coordenada en la que se inicializa el desvío.

direccion: La dirección del desvío.

id: El identificador del desvío.

pub fn inicializar_enemigo(&mut self, coordenada: Coordenada, vida: i8)

Inicializa un enemigo en el juego.

Esta función crea y agrega un nuevo enemigo al juego con la coordenada y puntos de vida especificados.

Argumentos

coordenada: La coordenada en la que se inicializa el enemigo.

vida: Los puntos de vida del enemigo.

pub fn inicializar_roca(&mut self, coordenada: Coordenada)

Inicializa una roca en el juego.

Esta función crea y agrega una nueva roca al juego en la coordenada especificada.

Argumentos

coordenada: La coordenada en la que se inicializa la roca.

pub fn inicializar_bomba( &mut self, coordenada: Coordenada, alcance: i8, tipo: TipoDeBomba, id: String )

Inicializa una bomba en el juego.

Esta función crea y agrega una nueva bomba al juego con la coordenada, alcance, tipo e identificador especificados.

Argumentos

coordenada: La coordenada en la que se inicializa la bomba.

alcance: El alcance de la bomba.

tipo: El tipo de bomba.

id: El identificador de la bomba.

pub fn inicializar_pared(&mut self, coordenada: Coordenada)

Inicializa una pared en el juego.

Esta función crea y agrega una nueva pared al juego en la coordenada especificada.

Argumentos

coordenada: La coordenada en la que se inicializa la pared.

fn posicionar_enemigos(&self, tablero: &mut [Vec<String>])

Posiciona los enemigos en el tablero del juego.

Esta función coloca los enemigos en las coordenadas correspondientes en el tablero, utilizando una representación única que combina el identificador y los puntos de vida del enemigo.

Argumentos

tablero: Una referencia mutable al tablero del juego representado como una matriz de cadenas.

fn posicionar_obstaculos(&self, tablero: &mut [Vec<String>])

Posiciona los obstáculos en el tablero del juego.

Esta función coloca los obstáculos en las coordenadas correspondientes en el tablero, utilizando el identificador del obstáculo como representación.

Argumentos

tablero: Una referencia mutable al tablero del juego representado como una matriz de cadenas.

fn posicionar_bombas(&self, tablero: &mut [Vec<String>])

Posiciona las bombas en el tablero del juego.

Esta función coloca las bombas en las coordenadas correspondientes en el tablero, utilizando el identificador de la bomba como representación.

Argumentos

tablero: Una referencia mutable al tablero del juego representado como una matriz de cadenas.

fn posicionar_desvios(&self, tablero: &mut [Vec<String>])

Posiciona los desvíos en el tablero del juego.

Esta función coloca los desvíos en las coordenadas correspondientes en el tablero, utilizando el identificador del desvío como representación.

Argumentos

tablero: Una referencia mutable al tablero del juego representado como una matriz de cadenas.

fn posicionar_elementos_en_tablero(&self) -> Vec<Vec<String>>

Posiciona todos los elementos en el tablero del juego.

Esta función crea un nuevo tablero y coloca en él todos los elementos del juego, incluyendo enemigos, obstáculos, bombas y desvíos. El tablero es representado como una matriz de cadenas.

Retorno

Un vector de vectores de cadenas que representa el tablero del juego.

pub fn imprimir_tablero(&self, tablero: &Vec<Vec<String>>)

Imprime el tablero en la consola.

Esta función imprime el contenido del tablero en la consola. Cada elemento del tablero se separa por espacio, y cada fila del tablero se imprime en una nueva línea.

Argumentos

tablero: Un vector de vectores de cadenas que representa el tablero del juego.

pub fn imprimir_tablero_en_archivo( &self, output_file: &mut File, tablero: &Vec<Vec<String>> ) -> Result<()>

Imprime el tablero en un archivo.

Esta función imprime el contenido del tablero en un archivo abierto en modo escritura. Cada elemento del tablero se separa por espacio, y cada fila del tablero se imprime en una nueva línea en el archivo.

Argumentos

output_file: Una referencia mutable a un archivo abierto en modo escritura.

tablero: Un vector de vectores de cadenas que representa el tablero del juego.

Errores

Esta función devuelve un error si no se puede escribir en el archivo.

pub fn realizar_jugada( &mut self, output_file: &mut File, coordenada: Coordenada ) -> Result<()>

Realiza una jugada en el juego y guarda el estado final en un archivo.

Esta función realiza una jugada en el juego, detonando una bomba en la coordenada especificada. Luego, muestra el tablero inicial y el tablero final en la consola, y guarda el tablero final en un archivo.

Argumentos

output_file: Una referencia mutable a un archivo abierto en modo escritura donde se guardará el tablero final.

coordenada: La coordenada en la que se detonará la bomba.

Errores

Esta función devuelve un error si no se puede escribir en el archivo de salida.

fn eliminar_enemigo( &mut self, coordenada: Coordenada, coordenada_bomba: Coordenada )

Elimina un enemigo del juego y actualiza su estado.

Esta función elimina un enemigo del juego si la coordenada especificada coincide con su posición. Si la bomba que causó la eliminación del enemigo ya se ha registrado previamente, no se realiza ninguna acción adicional. Si el enemigo aún tiene puntos de vida, se reduce su vida en uno y se actualiza su lista de bombas que lo impactaron. Si el enemigo queda con cero puntos de vida, se elimina completamente del juego.

Argumentos

coordenada: La coordenada en la que se encuentra el enemigo.

coordenada_bomba: La coordenada de la bomba que impactó al enemigo.

fn evaluar_casillero( &mut self, coordenada: &mut Coordenada, tablero: &mut Vec<Vec<String>>, i: i8, bomba: &Bomba, coordenada_original: Coordenada )

Evalúa el contenido de un casillero en el tablero y realiza las acciones correspondientes.

Esta función evalúa el contenido de un casillero en el tablero en función de su tipo y estado actual. Realiza acciones como eliminar enemigos, detonar bombas, desviar la trayectoria de una bomba o manejar obstáculos.

Argumentos

coordenada: Una referencia mutable a la coordenada actual en el tablero.

tablero: Una referencia mutable al tablero del juego.

i: Un contador que representa el alcance actual de la explosión de una bomba.

bomba: Una referencia a la bomba actual que está siendo evaluada.

coordenada_original: La coordenada original de la bomba antes de comenzar su trayectoria.

fn evaluar_arriba( &mut self, coordenada: &Coordenada, tablero: &mut Vec<Vec<String>>, i: i8, bomba: &Bomba )

Evalúa las posiciones hacia arriba desde la coordenada actual en busca de obstáculos o enemigos.

Esta función verifica las posiciones hacia arriba desde la coordenada actual y realiza las acciones correspondientes, como eliminar enemigos, detonar bombas o desviar la trayectoria de una bomba.

Argumentos

coordenada: Una referencia inmutable a la coordenada actual en el tablero.

tablero: Una referencia mutable al tablero del juego.

i: Un contador que representa el alcance actual de la explosión de una bomba.

bomba: Una referencia a la bomba actual que está siendo evaluada.

fn evaluar_abajo( &mut self, coordenada: &Coordenada, tablero: &mut Vec<Vec<String>>, i: i8, bomba: &Bomba )

Evalúa las posiciones hacia abajo desde la coordenada actual en busca de obstáculos o enemigos.

Esta función verifica las posiciones hacia abajo desde la coordenada actual y realiza las acciones correspondientes, como eliminar enemigos, detonar bombas o desviar la trayectoria de una bomba.

Argumentos

coordenada: Una referencia inmutable a la coordenada actual en el tablero.

tablero: Una referencia mutable al tablero del juego.

i: Un contador que representa el alcance actual de la explosión de una bomba.

bomba: Una referencia a la bomba actual que está siendo evaluada.

fn evaluar_izquierda( &mut self, coordenada: &Coordenada, tablero: &mut Vec<Vec<String>>, i: i8, bomba: &Bomba )

Evalúa las posiciones hacia la izquierda desde la coordenada actual en busca de obstáculos o enemigos.

Esta función verifica las posiciones hacia la izquierda desde la coordenada actual y realiza las acciones correspondientes, como eliminar enemigos, detonar bombas o desviar la trayectoria de una bomba.

Argumentos

coordenada: Una referencia inmutable a la coordenada actual en el tablero.

tablero: Una referencia mutable al tablero del juego.

i: Un contador que representa el alcance actual de la explosión de una bomba.

bomba: Una referencia a la bomba actual que está siendo evaluada.

fn evaluar_derecha( &mut self, coordenada: &Coordenada, tablero: &mut Vec<Vec<String>>, i: i8, bomba: &Bomba )

Evalúa las posiciones hacia la derecha desde la coordenada actual en busca de obstáculos o enemigos.

Esta función verifica las posiciones hacia la derecha desde la coordenada actual y realiza las acciones correspondientes, como eliminar enemigos, detonar bombas o desviar la trayectoria de una bomba.

Argumentos

coordenada: Una referencia inmutable a la coordenada actual en el tablero.

tablero: Una referencia mutable al tablero del juego.

i: Un contador que representa el alcance actual de la explosión de una bomba.

bomba: Una referencia a la bomba actual que está siendo evaluada.

fn funcion_bomba(&mut self, bomba: &Bomba, tablero: &mut Vec<Vec<String>>)

Evalúa el alcance de una bomba y su impacto en el tablero.

Esta función evalúa el alcance de una bomba y su impacto en el tablero, considerando las direcciones arriba, abajo, izquierda y derecha desde la posición de la bomba. Realiza las acciones correspondientes como eliminar enemigos, detonar bombas o desviar la trayectoria de una bomba en cada dirección.

Argumentos

bomba: Una referencia a la bomba que se va a evaluar.

tablero: Una referencia mutable al tablero del juego.

fn detonar_bomba( &mut self, tablero: &mut Vec<Vec<String>>, coordenada: Coordenada )

Detona una bomba en una coordenada dada y afecta el tablero.

Esta función busca una bomba en la coordenada especificada y la detona si aún no ha sido detonada. Cuando una bomba se detona, se evalúa su alcance en todas las direcciones y se aplican efectos como la eliminación de enemigos, la detonación de otras bombas cercanas o la desviación de trayectorias de bombas, afectando el tablero.

Argumentos

tablero: Una referencia mutable al tablero del juego.

coordenada: La coordenada en la que se va a detonar la bomba.