Struct tp0::juego::Juego

pub struct Juego {
    pub dimension: i8,
    pub enemigos: Vec<Enemigo>,
    /* private fields */
}

Fields
dimension: i8enemigos: Vec<Enemigo>
Implementations

impl Juego

pub fn new() -> Juego

pub fn inicializar_dimension(&mut self, dimension: i8)

Inicializa la dimensión del juego.

Esta función establece la dimensión del juego con el valor proporcionado.

Argumentos

dimension: El valor de la dimensión del juego.

pub fn inicializar_desvio( &mut self, coordenada: Coordenada, direccion: String, id: String )

Inicializa un desvío en el juego.

Esta función crea y agrega un nuevo desvío al juego con la coordenada, dirección e identificador especificados.

Argumentos

coordenada: La coordenada en la que se inicializa el desvío.

direccion: La dirección del desvío.

id: El identificador del desvío.

pub fn inicializar_enemigo(&mut self, coordenada: Coordenada, vida: i8)

Inicializa un enemigo en el juego.

Esta función crea y agrega un nuevo enemigo al juego con la coordenada y puntos de vida especificados.

Argumentos

coordenada: La coordenada en la que se inicializa el enemigo.

vida: Los puntos de vida del enemigo.

pub fn inicializar_roca(&mut self, coordenada: Coordenada)

Inicializa una roca en el juego.

Esta función crea y agrega una nueva roca al juego en la coordenada especificada.

Argumentos

coordenada: La coordenada en la que se inicializa la roca.

pub fn inicializar_bomba( &mut self, coordenada: Coordenada, alcance: i8, tipo: TipoDeBomba, id: String )

Inicializa una bomba en el juego.

Esta función crea y agrega una nueva bomba al juego con la coordenada, alcance, tipo e identificador especificados.

Argumentos

coordenada: La coordenada en la que se inicializa la bomba.

alcance: El alcance de la bomba.

tipo: El tipo de bomba.

id: El identificador de la bomba.

pub fn inicializar_pared(&mut self, coordenada: Coordenada)

Inicializa una pared en el juego.

Esta función crea y agrega una nueva pared al juego en la coordenada especificada.

Argumentos

coordenada: La coordenada en la que se inicializa la pared.

pub fn imprimir_tablero_en_archivo( &self, output_file: &mut File, tablero: &Vec<Vec<String>> ) -> Result<()>

Imprime el tablero en la consola.

Esta función imprime el contenido del tablero en la consola. Cada elemento del tablero se separa por espacio, y cada fila del tablero se imprime en una nueva línea.

Argumentos

tablero: Un vector de vectores de cadenas que representa el tablero del juego.

Imprime el tablero en un archivo.

Esta función imprime el contenido del tablero en un archivo abierto en modo escritura. Cada elemento del tablero se separa por espacio, y cada fila del tablero se imprime en una nueva línea en el archivo.

Argumentos

output_file: Una referencia mutable a un archivo abierto en modo escritura.

tablero: Un vector de vectores de cadenas que representa el tablero del juego.

Errores

Esta función devuelve un error si no se puede escribir en el archivo.

pub fn realizar_jugada( &mut self, output_file: &mut File, coordenada: Coordenada ) -> Result<()>

Realiza una jugada en el juego y guarda el estado final en un archivo.

Esta función realiza una jugada en el juego, detonando una bomba en la coordenada especificada. Luego, muestra el tablero inicial y el tablero final en la consola, y guarda el tablero final en un archivo.

Argumentos

output_file: Una referencia mutable a un archivo abierto en modo escritura donde se guardará el tablero final.

coordenada: La coordenada en la que se detonará la bomba.

Errores

Esta función devuelve un error si no se puede escribir en el archivo de salida.

pub fn buscar_enemigo(&self, coordenada: Coordenada) -> Option<usize>

Busca un enemigo en el vector de enemigos en función de sus coordenadas.

Argumentos

coordenada: La coordenada en la que se busca al enemigo.

Devolución

Si se encuentra un enemigo con las coordenadas especificadas, se devuelve Some(usize) que contiene la posición del enemigo en el vector de enemigos. Si no se encuentra ningún enemigo con las coordenadas dadas, se devuelve None.

pub fn eliminar_enemigo( &mut self, coordenada: Coordenada, coordenada_bomba: Coordenada )

Elimina un enemigo en función de sus coordenadas y la coordenada de la bomba.

Si se encuentra un enemigo en las coordenadas especificadas y la bomba no ha impactado previamente en él, reduce la vida del enemigo y actualiza su lista de bombas. Si la vida del enemigo llega a 0, se elimina del vector de enemigos.

Argumentos

coordenada: La coordenada en la que se busca al enemigo.

coordenada_bomba: La coordenada de la bomba.

pub fn evaluar_casillero( &mut self, coordenada: &mut Coordenada, tablero: &mut Vec<Vec<String>>, i: i8, bomba: &Bomba, coordenada_original: Coordenada )

Evalúa el contenido de un casillero en el tablero y realiza las acciones correspondientes.

Esta función evalúa el contenido de un casillero en el tablero en función de su tipo y estado actual. Realiza acciones como eliminar enemigos, detonar bombas, desviar la trayectoria de una bomba o manejar obstáculos.

Argumentos

coordenada: Una referencia mutable a la coordenada actual en el tablero.

tablero: Una referencia mutable al tablero del juego.

i: Un contador que representa el alcance actual de la explosión de una bomba.

bomba: Una referencia a la bomba actual que está siendo evaluada.

coordenada_original: La coordenada original de la bomba antes de comenzar su trayectoria.

pub fn evaluar_arriba( &mut self, coordenada: &Coordenada, tablero: &mut Vec<Vec<String>>, i: i8, bomba: &Bomba )

Evalúa las posiciones hacia arriba desde la coordenada actual en busca de obstáculos o enemigos.

Esta función verifica las posiciones hacia arriba desde la coordenada actual y realiza las acciones correspondientes, como eliminar enemigos, detonar bombas o desviar la trayectoria de una bomba.

Argumentos

coordenada: Una referencia inmutable a la coordenada actual en el tablero.

tablero: Una referencia mutable al tablero del juego.

i: Un contador que representa el alcance actual de la explosión de una bomba.

bomba: Una referencia a la bomba actual que está siendo evaluada.

pub fn detonar_bomba( &mut self, tablero: &mut Vec<Vec<String>>, coordenada: Coordenada )

Detona una bomba en el tablero en la coordenada especificada.

Esta función busca una bomba en la lista de bombas del objeto actual (self) que coincide con la coordenada especificada. Si se encuentra una bomba válida, la función la detona y ejecuta la función funcion_bomba con la bomba como argumento. Si la bomba ya ha sido detonada, la función no hace nada.

Argumentos

tablero: Un tablero mutable representado como un vector bidimensional de cadenas.

coordenada: La coordenada en la que se busca y detona la bomba.